refactor(editor): Extract action buttons and emoji picker into separate components
- Move EditorPaneActionButtons to dedicated component
- Move EditorPaneEmojiPicker to dedicated component
- Extract type definitions to editorPaneTypes.ts
- Simplify EditorPane component by removing UI markup duplication
